Move and remove HairdresserController
